Avoid monkey-patching and extend ApplicationRecord/Job instead
This works better with zeitwerk and autoloading: stuff in the lib/ directory isn’t auto(re)loaded, which meant we had to manually require the path in the core_extensions initializer. We could have moved the extensions in /app or added /lib to the autoloaded paths, but even then, zeitwerk really doesn’t like autoloaded stuff in the initializers (basically it breaks reloading.)

Instead, explicitly including/extending modules in our ApplicationRecord and ApplicationJob “root” classes has the same net effect but autoloading works and no manual requires are needed. Yay!
