+++ app/jobs/application_job.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 class ApplicationJob < ActiveJob::Base
+  extend JobExtensions::RemoveDelayedJobs
 end
